// types/visits.ts
export interface PaidProduct {
  id: string;
  company_id: string;
  name: string;
  unit_type?: string | null;
  unit?: string | null;
  price: number;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}